+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## Unreleased
+
+### Added
+- [#333](https://github.com/cloudcreativity/laravel-json-api/issues/333)
+Eloquent adapters now have a `filterWithScopes()` method, that maps JSON API filters to
+model scopes and the Eloquent `where*` method names. This is opt-in: i.e. to use, the
+developer must call `filterWithScopes()` within their adapter's `filter()` method.
+
 ## [1.3.0] - 2019-07-24
 
 Package now supports Laravel 5.9.